requirements.txt
Django==4.2.13
pillow==10.4.0
django-admin-sortable2==2.2.3
django-tinymce==4.1.0
environs==11.0.0
requests==2.32.3
